requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
gunicorn>=20.1
psycopg2-binary>=2.9
django-cors-headers>=3.14
python-dotenv>=1.0